# backend/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/student/profile/schooling": {
            "put": {
                "tags": [
                    "student-profile-controller"
                ],
                "operationId": "updateSchoolingDetails",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SchoolingDetailsDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/profile/personal": {
            "put": {
                "tags": [
                    "student-profile-controller"
                ],
                "operationId": "updatePersonalDetails",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonalDetailsDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/profile/contact": {
            "put": {
                "tags": [
                    "student-profile-controller"
                ],
                "operationId": "updateContactDetails",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactDetailsDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/profile/academic": {
            "put": {
                "tags": [
                    "student-profile-controller"
                ],
                "operationId": "updateAcademicRecord",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AcademicRecordDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/student/profile": {
            "get": {
                "tags": [
                    "student-profile-controller"
                ],
                "operationId": "getMyProfile",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseStudentProfileDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/profile/verification-status": {
            "get": {
                "tags": [
                    "student-profile-controller"
                ],
                "operationId": "getVerificationStatus",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseString"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SchoolingDetailsDto": {
                "type": "object",
                "properties": {
                    "xiiMarksPercentage": {
                        "type": "number",
                        "format": "double"
                    },
                    "xmarksPercentage": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PersonalDetailsDto": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date"
                    },
                    "bio": {
                        "type": "string"
                    }
                }
            },
            "ContactDetailsDto": {
                "type": "object",
                "properties": {
                    "alternateEmail": {
                        "type": "string"
                    },
                    "studentMobile1": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    }
                }
            },
            "AcademicRecordDto": {
                "type": "object",
                "properties": {
                    "cgpa": {
                        "type": "number",
                        "format": "double"
                    },
                    "standingArrears": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "historyOfArrears": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "ugYearOfPass": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "ApiResponseString": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ApiResponseStudentProfileDto": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/StudentProfileDto"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "StudentProfileDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "rollNo": {
                        "type": "string"
                    },
                    "batch": {
                        "type": "string"
                    },
                    "resumeUrl": {
                        "type": "string"
                    },
                    "verificationStatus": {
                        "type": "string"
                    },
                    "isLocked": {
                        "type": "boolean"
                    },
                    "isEligibleForPlacements": {
                        "type": "boolean"
                    },
                    "interestedOnPlacement": {
                        "type": "boolean"
                    },
                    "isPlaced": {
                        "type": "boolean"
                    },
                    "numberOfOffers": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "highestPackageLpa": {
                        "type": "number",
                        "format": "double"
                    },
                    "personalDetails": {
                        "$ref": "#/components/schemas/PersonalDetailsDto"
                    },
                    "contactDetails": {
                        "$ref": "#/components/schemas/ContactDetailsDto"
                    },
                    "academicRecord": {
                        "$ref": "#/components/schemas/AcademicRecordDto"
                    },
                    "schoolingDetails": {
                        "$ref": "#/components/schemas/SchoolingDetailsDto"
                    }
                }
            },
...
        }
    }
}